+++ test/Microsoft.AspNetCore.Mvc.Core.Test/AcceptedResultTests.cs
@@ -141,7 +141,7 @@ namespace Microsoft.AspNetCore.Mvc.Core.Test
             options.Value.OutputFormatters.Add(formatter.Object);
             var services = new ServiceCollection();
             services.AddSingleton<IActionResultExecutor<ObjectResult>>(new ObjectResultExecutor(
-                options,
+                new DefaultOutputFormatterSelector(NullLoggerFactory.Instance, options),
                 new TestHttpResponseStreamWriterFactory(),
                 NullLoggerFactory.Instance));
