+++ test/Microsoft.AspNetCore.Mvc.Core.Test/HttpOkObjectResultTest.cs
@@ -79,7 +79,7 @@ namespace Microsoft.AspNetCore.Mvc
 
             var services = new ServiceCollection();
             services.AddSingleton<IActionResultExecutor<ObjectResult>>(new ObjectResultExecutor(
-                options,
+                new DefaultOutputFormatterSelector(NullLoggerFactory.Instance, options),
                 new TestHttpResponseStreamWriterFactory(),
                 NullLoggerFactory.Instance));
