+++ test/Microsoft.AspNetCore.Mvc.Core.Test/Infrastructure/ObjectResultExecutorTest.cs
@@ -21,48 +21,19 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
 {
     public class ObjectResultExecutorTest
     {
-        [Fact]
-        public void SelectFormatter_WithNoProvidedContentType_DoesConneg()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new TestXmlOutputFormatter(),
-                new TestJsonOutputFormatter(), // This will be chosen based on the accept header
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "application/json";
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { "application/json" },
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[1], formatter);
-            MediaTypeAssert.Equal("application/json", context.ContentType);
-        }
-
         // For this test case probably the most common use case is when there is a format mapping based
         // content type selected but the developer had set the content type on the Response.ContentType
         [Fact]
         public async Task ExecuteAsync_ContentTypeProvidedFromResponseAndObjectResult_UsesResponseContentType()
         {
             // Arrange
-            var executor = CreateCustomObjectResultExecutor();
+            var executor = CreateExecutor();
+
             var httpContext = new DefaultHttpContext();
             var actionContext = new ActionContext() { HttpContext = httpContext };
             httpContext.Request.Headers[HeaderNames.Accept] = "application/xml"; // This will not be used
             httpContext.Response.ContentType = "text/plain";
+
             var result = new ObjectResult("input");
             result.Formatters.Add(new TestXmlOutputFormatter());
             result.Formatters.Add(new TestJsonOutputFormatter());
@@ -72,50 +43,20 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
             await executor.ExecuteAsync(actionContext, result);
 
             // Assert
-            Assert.IsType<TestStringOutputFormatter>(executor.SelectedOutputFormatter);
             MediaTypeAssert.Equal("text/plain; charset=utf-8", httpContext.Response.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithOneProvidedContentType_IgnoresAcceptHeader()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new TestXmlOutputFormatter(),
-                new TestJsonOutputFormatter(), // This will be chosen based on the content type
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "application/xml"; // This will not be used
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { "application/json" },
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[1], formatter);
-            Assert.Equal(new StringSegment("application/json"), context.ContentType);
         }
 
         [Fact]
         public async Task ExecuteAsync_WithOneProvidedContentType_FromResponseContentType_IgnoresAcceptHeader()
         {
             // Arrange
-            var executor = CreateCustomObjectResultExecutor();
+            var executor = CreateExecutor();
+
             var httpContext = new DefaultHttpContext();
             var actionContext = new ActionContext() { HttpContext = httpContext };
             httpContext.Request.Headers[HeaderNames.Accept] = "application/xml"; // This will not be used
             httpContext.Response.ContentType = "application/json";
+
             var result = new ObjectResult("input");
             result.Formatters.Add(new TestXmlOutputFormatter());
             result.Formatters.Add(new TestJsonOutputFormatter()); // This will be chosen based on the content type
@@ -124,48 +65,20 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
             await executor.ExecuteAsync(actionContext, result);
 
             // Assert
-            Assert.IsType<TestJsonOutputFormatter>(executor.SelectedOutputFormatter);
             Assert.Equal("application/json; charset=utf-8", httpContext.Response.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithOneProvidedContentType_NoFallback()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new TestXmlOutputFormatter(),
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "application/xml"; // This will not be used
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { "application/json" },
-                formatters);
-
-            // Assert
-            Assert.Null(formatter);
         }
 
         [Fact]
         public async Task ExecuteAsync_WithOneProvidedContentType_FromResponseContentType_NoFallback()
         {
             // Arrange
-            var executor = CreateCustomObjectResultExecutor();
+            var executor = CreateExecutor();
+
             var httpContext = new DefaultHttpContext();
             var actionContext = new ActionContext() { HttpContext = httpContext };
             httpContext.Request.Headers[HeaderNames.Accept] = "application/xml"; // This will not be used
             httpContext.Response.ContentType = "application/json";
+
             var result = new ObjectResult("input");
             result.Formatters.Add(new TestXmlOutputFormatter());
 
@@ -173,268 +86,7 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
             await executor.ExecuteAsync(actionContext, result);
 
             // Assert
-            Assert.Null(executor.SelectedOutputFormatter);
-        }
-
-        // ObjectResult.ContentTypes, Accept header, expected content type
-        public static TheoryData<MediaTypeCollection, string, string> ContentTypes
-        {
-            get
-            {
-                var contentTypes = new MediaTypeCollection
-                {
-                    "text/plain",
-                    "text/xml",
-                    "application/json",
-                };
-
-                return new TheoryData<MediaTypeCollection, string, string>()
-                {
-                    // Empty accept header, should select based on ObjectResult.ContentTypes.
-                    { contentTypes, "", "application/json" },
-
-                    // null accept header, should select based on ObjectResult.ContentTypes.
-                    { contentTypes, null, "application/json" },
-
-                    // The accept header does not match anything in ObjectResult.ContentTypes.
-                    // The first formatter that can write the result gets to choose the content type.
-                    { contentTypes, "text/custom", "application/json" },
-
-                    // Accept header matches ObjectResult.ContentTypes, but no formatter supports the accept header.
-                    // The first formatter that can write the result gets to choose the content type.
-                    { contentTypes, "text/xml", "application/json" },
-
-                    // Filters out Accept headers with 0 quality and selects the one with highest quality.
-                    {
-                        contentTypes,
-                        "text/plain;q=0.3, text/json;q=0, text/cusotm;q=0.0, application/json;q=0.4",
-                        "application/json"
-                    },
-                };
-            }
-        }
-
-        [Theory]
-        [MemberData(nameof(ContentTypes))]
-        public void SelectFormatter_WithMultipleProvidedContentTypes_DoesConneg(
-            MediaTypeCollection contentTypes,
-            string acceptHeader,
-            string expectedContentType)
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new CannotWriteFormatter(),
-                new TestJsonOutputFormatter(),
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = acceptHeader;
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                contentTypes,
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[1], formatter);
-            Assert.Equal(new StringSegment(expectedContentType), context.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_NoProvidedContentTypesAndNoAcceptHeader_ChoosesFirstFormatterThatCanWrite()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new CannotWriteFormatter(),
-                new TestJsonOutputFormatter(),
-                new TestXmlOutputFormatter(),
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection(),
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[1], formatter);
-            Assert.Equal(new StringSegment("application/json"), context.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithAcceptHeader_UsesFallback()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new TestXmlOutputFormatter(),
-                new TestJsonOutputFormatter(),
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "text/custom,application/custom";
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { },
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[0], formatter);
-            Assert.Equal(new StringSegment("application/xml"), context.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithAcceptHeaderAndReturnHttpNotAcceptable_DoesNotUseFallback()
-        {
-            // Arrange
-            var options = Options.Create(new MvcOptions());
-            options.Value.ReturnHttpNotAcceptable = true;
-
-            var executor = CreateExecutor(options);
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new TestXmlOutputFormatter(),
-                new TestJsonOutputFormatter(),
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "text/custom,application/custom";
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { },
-                formatters);
-
-            // Assert
-            Assert.Null(formatter);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithAcceptHeaderOnly_SetsContentTypeIsServerDefinedToFalse()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new ServerContentTypeOnlyFormatter()
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "text/custom";
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                new MediaTypeCollection { },
-                formatters);
-
-            // Assert
-            Assert.Null(formatter);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithAcceptHeaderAndContentTypes_SetsContentTypeIsServerDefinedWhenExpected()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new ServerContentTypeOnlyFormatter()
-            };
-            
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            context.HttpContext.Request.Headers[HeaderNames.Accept] = "text/custom, text/custom2";
-
-            var serverDefinedContentTypes = new MediaTypeCollection();
-            serverDefinedContentTypes.Add("text/other");
-            serverDefinedContentTypes.Add("text/custom2");
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                serverDefinedContentTypes,
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[0], formatter);
-            Assert.Equal(new StringSegment("text/custom2"), context.ContentType);
-        }
-
-        [Fact]
-        public void SelectFormatter_WithContentTypesOnly_SetsContentTypeIsServerDefinedToTrue()
-        {
-            // Arrange
-            var executor = CreateExecutor();
-
-            var formatters = new List<IOutputFormatter>
-            {
-                new ServerContentTypeOnlyFormatter()
-            };
-
-            var context = new OutputFormatterWriteContext(
-                new DefaultHttpContext(),
-                new TestHttpResponseStreamWriterFactory().CreateWriter,
-                objectType: null,
-                @object: null);
-
-            var serverDefinedContentTypes = new MediaTypeCollection();
-            serverDefinedContentTypes.Add("text/custom");
-
-            // Act
-            var formatter = executor.SelectFormatter(
-                context,
-                serverDefinedContentTypes,
-                formatters);
-
-            // Assert
-            Assert.Same(formatters[0], formatter);
-            Assert.Equal(new StringSegment("text/custom"), context.ContentType);
+            Assert.Equal(406, httpContext.Response.StatusCode);
         }
 
         [Fact]
@@ -637,20 +289,10 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
             return httpContext;
         }
 
-        private static TestObjectResultExecutor CreateExecutor(IOptions<MvcOptions> options = null)
+        private static ObjectResultExecutor CreateExecutor(IOptions<MvcOptions> options = null)
         {
-            return new TestObjectResultExecutor(
-                options ?? Options.Create(new MvcOptions()),
-                new TestHttpResponseStreamWriterFactory(),
-                NullLoggerFactory.Instance);
-        }
-
-        private static CustomObjectResultExecutor CreateCustomObjectResultExecutor()
-        {
-            return new CustomObjectResultExecutor(
-                Options.Create(new MvcOptions()),
-                new TestHttpResponseStreamWriterFactory(),
-                NullLoggerFactory.Instance);
+            var selector = new DefaultOutputFormatterSelector(NullLoggerFactory.Instance, options ?? Options.Create<MvcOptions>(new MvcOptions()));
+            return new ObjectResultExecutor(selector, new TestHttpResponseStreamWriterFactory(), NullLoggerFactory.Instance);
         }
 
         private class CannotWriteFormatter : IOutputFormatter
@@ -710,47 +352,6 @@ namespace Microsoft.AspNetCore.Mvc.Infrastructure
             public override Task WriteResponseBodyAsync(OutputFormatterWriteContext context, Encoding selectedEncoding)
             {
                 return Task.FromResult(0);
-            }
-        }
-
-        private class TestObjectResultExecutor : ObjectResultExecutor
-        {
-            public TestObjectResultExecutor(
-                IOptions<MvcOptions> options,
-                IHttpResponseStreamWriterFactory writerFactory,
-                ILoggerFactory loggerFactory)
-                : base(options, writerFactory, loggerFactory)
-            {
-            }
-
-            new public IOutputFormatter SelectFormatter(
-                OutputFormatterWriteContext formatterContext,
-                MediaTypeCollection contentTypes,
-                IList<IOutputFormatter> formatters)
-            {
-                return base.SelectFormatter(formatterContext, contentTypes, formatters);
-            }
-        }
-
-        private class CustomObjectResultExecutor : ObjectResultExecutor
-        {
-            public CustomObjectResultExecutor(
-                IOptions<MvcOptions> options,
-                IHttpResponseStreamWriterFactory writerFactory,
-                ILoggerFactory loggerFactory)
-                : base(options, writerFactory, loggerFactory)
-            {
-            }
-
-            public IOutputFormatter SelectedOutputFormatter { get; private set; }
-
-            protected override IOutputFormatter SelectFormatter(
-                OutputFormatterWriteContext formatterContext,
-                MediaTypeCollection contentTypes,
-                IList<IOutputFormatter> formatters)
-            {
-                SelectedOutputFormatter = base.SelectFormatter(formatterContext, contentTypes, formatters);
-                return SelectedOutputFormatter;
             }
         }
