+++ test/Microsoft.AspNetCore.Mvc.Core.Test/Internal/ControllerActionInvokerTest.cs
@@ -1596,14 +1596,14 @@ namespace Microsoft.AspNetCore.Mvc.Internal
             }
 
             var httpContext = new DefaultHttpContext();
-            var options = new MvcOptions();
-            var mvcOptionsAccessor = Options.Create(options);
+
+            var options = Options.Create(new MvcOptions());
 
             var services = new ServiceCollection();
             services.AddSingleton<ILoggerFactory>(NullLoggerFactory.Instance);
-            services.AddSingleton<IOptions<MvcOptions>>(mvcOptionsAccessor);
+            services.AddSingleton<IOptions<MvcOptions>>(options);
             services.AddSingleton<IActionResultExecutor<ObjectResult>>(new ObjectResultExecutor(
-                mvcOptionsAccessor,
+                new DefaultOutputFormatterSelector(NullLoggerFactory.Instance, options),
                 new TestHttpResponseStreamWriterFactory(),
                 NullLoggerFactory.Instance));
 
@@ -1622,7 +1622,7 @@ namespace Microsoft.AspNetCore.Mvc.Internal
                     await c.HttpContext.Response.WriteAsync(c.Object.ToString());
                 });
 
-            options.OutputFormatters.Add(formatter.Object);
+            options.Value.OutputFormatters.Add(formatter.Object);
 
             var diagnosticSource = new DiagnosticListener("Microsoft.AspNetCore");
             if (diagnosticListener != null)
