+++ test/Microsoft.AspNetCore.Mvc.WebApiCompatShimTest/ExceptionResultTest.cs
@@ -76,7 +76,7 @@ namespace System.Web.Http
 
             var services = new ServiceCollection();
             services.AddSingleton<IActionResultExecutor<ObjectResult>>(new ObjectResultExecutor(
-                options,
+                new DefaultOutputFormatterSelector(NullLoggerFactory.Instance, options),
                 new TestHttpResponseStreamWriterFactory(),
                 NullLoggerFactory.Instance));
